+++ core/sonic/src/helped/helper.rs
@@ -15,7 +15,7 @@
 
 use pairing::{Engine, CurveAffine, CurveProjective, Field};
 use crate::srs::SRS;
-use crate::polynomials::commitment::multiexp;
+use crate::utils::multiexp;
 
 #[derive(Clone)]
 pub struct Batch<E: Engine> {
