+++ core/sonic/src/unhelped/mod.rs
@@ -1,0 +1,3 @@
+pub mod grand_product;
+pub mod permutation;
+pub mod well_formed;
